refactor(models): split models.rs into per-domain submodules
`models.rs` had grown to 2418 lines mixing disc, title, menu, asset, and
build-settings schemas in one file (#48). Split into `models/{disc,title,
menu,asset,build_settings,render_target}.rs`, with cross-domain shared types
(`PlaybackAction`, `AspectMode`) and the project-level container types kept
in `models/mod.rs` alongside the existing integration-style test suite.

Purely organisational: `lib.rs`'s `pub use models::*;` already flattened
everything onto the crate root, and every consumer already imports via
`crate::models::{...}` or `crate::models::*`, so no call sites change.
`models/mod.rs` re-exports each submodule with `pub use` to preserve that.

Verified with `cargo build --workspace`, `cargo test --workspace` (189
passing), `cargo clippy --workspace --all-targets -- -D warnings`, and
`cargo fmt --check`.
